requirements.txt
aiofiles
config
environ
fastapi
fastapi-authz
fastapi_login
fastapi-sessions
pydantic
psycopg2
pytest
python-decouple
python-dotenv
python-multipart
pyotp
sendgrid
sqlalchemy
uvicorn
twilio